tests/wpt/tests/html/canvas/offscreen/canvas-host/2d.canvas.host.size.attributes.parse.hex.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.canvas.host.size.attributes.parse.hex</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.canvas.host.size.attributes.parse.hex</h1>
<p class="desc">Parsing of non-negative integers</p>


<script>
var t = async_test("Parsing of non-negative integers");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  canvas.width = '0x100';
  canvas.height = '0x100';
  _assertSame(canvas.width, 256, "canvas.width", "256");
  _assertSame(canvas.height, 256, "canvas.height", "256");
  t.done();

});
</script>
